Now there is finally a functionnal calculator
This calculator can :
Exectute mathematical operations (+,-,*,/,//,%,**)
Verify if user typed impossible events (e g  : division by 0) and reexecute the function if so
Verify also if user typed an existential operator (e g : not "m" of ","), and reexecute the function if so.
The code runs under a while loop where the function is being called all the time so the user can exectute as many operations as possible!

There is really, really, really, really more to go !